Improve tests' structure
Use a decorator so repeating of tests for JSON & gRPC channel isn't
required.
